Share header-env parsers, harden parseFeedbackExtraction, drop AbortController boilerplate
